requirements.txt
# Discord Bot
discord.py

# Environment variables
python-dotenv

# YouTube APIs
youtube-transcript-api
google-api-python-client
isodate

# Google Gemini API
google-generativeai

# Google Text-to-Speech
gTTS
